+++ zcash_client_backend/src/data_api/testing.rs
@@ -74,9 +74,12 @@ use crate::{
 
 #[cfg(feature = "transparent-inputs")]
 use {
-    super::{TransactionsInvolvingAddress, wallet::input_selection::ShieldingSelector},
-    crate::{data_api::Balance, wallet::TransparentAddressMetadata},
-    ::transparent::{address::TransparentAddress, keys::TransparentKeyScope},
+    super::{
+        TransactionsInvolvingAddress, TransparentBalances,
+        wallet::input_selection::ShieldingSelector,
+    },
+    crate::wallet::TransparentAddressMetadata,
+    ::transparent::address::TransparentAddress,
     zcash_keys::keys::transparent::gap_limits::GapLimits,
 };
 
@@ -2776,7 +2779,7 @@ impl WalletRead for MockWalletDb {
         _account: Self::AccountId,
         _target_height: TargetHeight,
         _confirmations_policy: ConfirmationsPolicy,
-    ) -> Result<HashMap<TransparentAddress, (TransparentKeyScope, Balance)>, Self::Error> {
+    ) -> Result<TransparentBalances, Self::Error> {
         Ok(HashMap::new())
     }